+++ branches/1.1/src/animators/AnimatorRotator.cpp
@@ -80,7 +80,7 @@ namespace aprilui
 			}
 			float oldAngle = angle;
 			angle += k * mSpeed;
-			int circles = floor(angle / 360);
+			int circles = (int)floor(angle / 360);
 			if ((mDestAngle >= 0.0f && hsgn(oldAngle + 360 * circles - mDestAngle) != hsgn(angle + 360 * circles - mDestAngle)) ||
 				(hsgn(oldAngle - 360 * circles - mDestAngle) != hsgn(angle - 360 * circles - mDestAngle)))
 			{
